+++ src/app/admin/page.tsx
@@ -1,15 +1,27 @@
 'use client'
 
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 import { Tabs, Tab, Stack } from '@mui/material'
-import Announce from './announce/page'
-import Tag from './tag/page'
+import Announce from './announce/Announce'
+import Tag from './tag/Tag'
+import Banner from './banner/Banner'
+import useAdminStore from '@/states/useAdminStore'
+import { useRouter } from 'next/navigation'
 const AdminLayout = () => {
   const [value, setValue] = useState(0)
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     setValue(newValue)
   }
+  const { isLoggedIn } = useAdminStore()
+  const router = useRouter()
+
+  useEffect(() => {
+    if (isLoggedIn === false) {
+      alert('로그인이 필요한 서비스입니다.')
+      router.push('/admin/login')
+    }
+  }, [isLoggedIn])
 
   return (
     <Stack direction={'row'}>
@@ -21,9 +33,11 @@ const AdminLayout = () => {
       >
         <Tab label="Announce" />
         <Tab label="Tag" />
+        <Tab label="Banner" />
       </Tabs>
       {value === 0 && <Announce />}
       {value === 1 && <Tag />}
+      {value === 2 && <Banner />}
     </Stack>
   )
 }
